requirements.txt
openai
deepgram-sdk
python-dotenv
sounddevice
soundfile
fastapi
uvicorn[standard]
python-multipart

